feat: general refactor done
