+++ lib/features/finance/modules/budget/presentation/sheets/edit_category_sheet.dart
@@ -1,0 +1,181 @@
+import 'package:flutter/material.dart';
+import 'package:keep_track/core/theme/app_theme.dart';
+import 'package:keep_track/features/finance/presentation/state/finance_category_controller.dart';
+
+import '../controllers/budget_controller.dart';
+import '../../domain/entities/budget.dart';
+import '../../domain/entities/budget_category.dart';
+
+class EditCategorySheet extends StatefulWidget {
+  final Budget group;
+  final BudgetCategory category;
+  final FinanceCategoryController categoryController;
+  final BudgetController budgetController;
+  final VoidCallback? onDelete;
+
+  const EditCategorySheet({
+    super.key,
+    required this.group,
+    required this.category,
+    required this.categoryController,
+    required this.budgetController,
+    this.onDelete,
+  });
+
+  @override
+  State<EditCategorySheet> createState() => _EditCategorySheetState();
+}
+
+class _EditCategorySheetState extends State<EditCategorySheet> {
+  late final TextEditingController _nameCtrl;
+  late final TextEditingController _amountCtrl;
+  bool _saving = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _nameCtrl = TextEditingController(
+      text: widget.category.financeCategory?.name ?? '',
+    );
+    _amountCtrl = TextEditingController(
+      text: widget.category.targetAmount.toStringAsFixed(2),
+    );
+  }
+
+  @override
+  void dispose() {
+    _nameCtrl.dispose();
+    _amountCtrl.dispose();
+    super.dispose();
+  }
+
+  Future<void> _save() async {
+    final name = _nameCtrl.text.trim();
+    final amount = double.tryParse(_amountCtrl.text);
+    if (name.isEmpty) {
+      ScaffoldMessenger.of(
+        context,
+      ).showSnackBar(const SnackBar(content: Text('Enter a category name')));
+      return;
+    }
+    if (amount == null || amount < 0) {
+      ScaffoldMessenger.of(
+        context,
+      ).showSnackBar(const SnackBar(content: Text('Enter a valid amount')));
+      return;
+    }
+    setState(() => _saving = true);
+    try {
+      // Update FinanceCategory name if changed
+      final fc = widget.category.financeCategory;
+      if (fc != null && fc.name != name) {
+        await widget.categoryController.updateCategory(fc.copyWith(name: name));
+      }
+      // Update planned amount if changed
+      if (amount != widget.category.targetAmount) {
+        await widget.budgetController.updateCategory(
+          widget.group.id!,
+          widget.category.copyWith(targetAmount: amount),
+        );
+      }
+      if (mounted) Navigator.pop(context);
+    } catch (e) {
+      if (mounted) {
+        ScaffoldMessenger.of(
+          context,
+        ).showSnackBar(SnackBar(content: Text('Error: $e')));
+        setState(() => _saving = false);
+      }
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: EdgeInsets.only(
+        bottom: MediaQuery.of(context).viewInsets.bottom,
+      ),
+      child: Container(
+        padding: const EdgeInsets.fromLTRB(24, 20, 24, 24),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Center(
+              child: Container(
+                width: 36,
+                height: 4,
+                decoration: BoxDecoration(
+                  color: AppColors.textTertiary.withValues(alpha: 0.3),
+                  borderRadius: BorderRadius.circular(2),
+                ),
+              ),
+            ),
+            const SizedBox(height: 16),
+            Text('Edit Category', style: AppTextStyles.h4),
+            const SizedBox(height: 4),
+            Text(
+              widget.group.title ?? '',
+              style: AppTextStyles.caption.copyWith(
+                color: AppColors.textSecondary,
+              ),
+            ),
+            const SizedBox(height: 20),
+            TextField(
+              controller: _nameCtrl,
+              autofocus: true,
+              textCapitalization: TextCapitalization.words,
+              decoration: const InputDecoration(
+                labelText: 'Category Name',
+                border: OutlineInputBorder(),
+              ),
+            ),
+            const SizedBox(height: 16),
+            TextField(
+              controller: _amountCtrl,
+              keyboardType: const TextInputType.numberWithOptions(
+                decimal: true,
+              ),
+              decoration: const InputDecoration(
+                labelText: 'Planned Amount',
+                border: OutlineInputBorder(),
+                prefixText: '₱ ',
+              ),
+              onSubmitted: (_) => _save(),
+            ),
+            const SizedBox(height: 24),
+            SizedBox(
+              width: double.infinity,
+              child: FilledButton(
+                onPressed: _saving ? null : _save,
+                child: _saving
+                    ? const SizedBox(
+                        width: 20,
+                        height: 20,
+                        child: CircularProgressIndicator(
+                          strokeWidth: 2,
+                          color: Colors.white,
+                        ),
+                      )
+                    : const Text('Save'),
+              ),
+            ),
+            if (widget.onDelete != null) ...[
+              const SizedBox(height: 8),
+              SizedBox(
+                width: double.infinity,
+                child: TextButton(
+                  onPressed: _saving ? null : widget.onDelete,
+                  child: const Text(
+                    'Delete Category',
+                    style: TextStyle(color: AppColors.error),
+                  ),
+                ),
+              ),
+            ],
+          ],
+        ),
+      ),
+    );
+  }
+}
