+++ lib/features/finance/modules/budget/presentation/sheets/start_planning_sheet.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:keep_track/core/theme/app_theme.dart';
 
 import '../../../../../../shared/infrastructure/supabase/supabase_service.dart';
-import '../../../../presentation/state/budget_controller.dart';
+import '../controllers/budget_controller.dart';
 import '../../../../presentation/state/month_plan_controller.dart';
-import '../screens/create_group_sheet.dart';
+import 'create_group_sheet.dart';
 
 class StartPlanningSheet extends StatefulWidget {
   final String monthKey;
@@ -17,6 +17,7 @@ class StartPlanningSheet extends StatefulWidget {
   final SupabaseService supabaseService;
 
   const StartPlanningSheet({
+    super.key,
     required this.monthKey,
     required this.monthLabel,
     required this.prevMonthKey,
