+++ lib/features/finance/modules/budget/presentation/widgets/category_row.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:keep_track/core/theme/app_theme.dart';
 
 import '../../domain/entities/budget_category.dart';
-import '../utils/currency_formatter.dart';
+import '../helpers/currency_formatter.dart';
 
 class CategoryRow extends StatefulWidget {
   final BudgetCategory category;
