+++ lib/features/finance/modules/budget/presentation/widgets/debt_detail_content.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:keep_track/core/theme/app_theme.dart';
-import 'package:keep_track/features/finance/modules/budget/presentation/utils/currency_formatter.dart';
+import 'package:keep_track/features/finance/modules/budget/presentation/helpers/currency_formatter.dart';
 
 import '../../../debt/domain/entities/debt.dart';
 import '../../../transaction/domain/entities/transaction.dart';
