+++ lib/features/finance/presentation/screens/finance_main_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:keep_track/core/ui/app_layout_controller.dart';
 import 'package:keep_track/core/ui/ui.dart';
 import 'tabs/accounts/accounts_tab_new.dart';
-import 'tabs/budgets/budgets_tab_new.dart';
+import '../../modules/budget/presentation/screens/budgets_tab_new.dart';
 import 'tabs/debts/debts_tab_new.dart';
 import 'tabs/goals/goals_tab.dart';
 import 'tabs/planned_payments/planned_payments_tab.dart';
